requirements.txt
python-dotenv==0.10.5
selenium==3.141.0
urllib3==1.25.8
